Add Additional jenkins container builds
There will be some TODO work here:
The rh-maven33 packages don't exist for Fedora, so the
version of maven will be different, and may not include
all packages.

The nodejs packages are similarly different versions, as is
the nodemon package which is an npm package for Fedora.

This PR is a first go at getting the containers to build.